+++ lib/modules/login/domain/entities/user.dart
@@ -1,0 +1,15 @@
+class User {
+  final String id;
+  final String name;
+  final String email;
+  final String? profileImageUrl;
+
+  User({
+    required this.id,
+    required this.name,
+    required this.email,
+    this.profileImageUrl,
+  });
+
+  User.empty() : id = '', name = '', email = '', profileImageUrl = null;
+}